[FIX] 장바구니 상품 이름 안보이는 문제 해결

+++ src/main/java/com/example/somsom_market/domain/CartTest/CartItemTest.java
@@ -32,7 +32,7 @@ public class CartItemTest implements Serializable {
         this.quantity = quantity;
     }
 
-    public double getTotalPrice() {
+    public int getTotalPrice() {
         if (item != null) {
             return item.getPrice() * quantity;
         }
